pyproject.toml
[project]
name = "auth0-test-streamlit-app"
version = "0.1.0"
description = "A Streamlit app to test Auth0 login flow and JWT handling"
requires-python = ">=3.12"
dependencies = [
    "streamlit>=1.32.0",
    "python-jose[cryptography]>=3.3.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "authlib>=1.3.0",
    "pyjwt>=2.10.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py312" 